packages/justhtml/justhtml-1.4.0.tar.gz/justhtml-1.4.0/tests/justhtml-tests/treebuilder_coverage.dat

#data
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//" "http://www.w3.org/TR/html4/loose.dtd">
#errors
(1,100): unknown-doctype

#document
| <!DOCTYPE html "-//W3C//DTD HTML 4.01 Transitional//" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|   <body>


#data
<math><mi>\x0ctext</mi></math>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         " text"


#data
<head></head></template>text
#errors
(1,1): expected-doctype-but-got-start-tag
(1,24): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     "text"


#data
<dl><dd>x</dd></dl>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "x"


#data
<table><caption><!--comment--></caption></table>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <!-- comment -->


#data
<select>\x0ctext</select>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <select>
|       " text"


#data
<frameset></frameset></frameset>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,32): unexpected-token-after-frameset

#document
| <html>
|   <head>
|   <frameset>



#data
<table><caption><table></table></caption></table>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,23): unexpected-start-tag-implies-end-tag
(1,23): unexpected-start-tag-implies-end-tag
(1,41): unexpected-end-tag
(1,49): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|     <table>


#data
<table>\x00</table>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,8): unexpected-null-character

#document
| <html>
|   <head>
|   <body>
|     <table>


#data
<table>text</table>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,11): unexpected-character-implies-table-voodoo

#document
| <html>
|   <head>
|   <body>
|     "text"
|     <table>


#data
<table><td><!-- comment in cell --></td></table>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,11): unexpected-cell-in-table-body

#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <!--  comment in cell  -->


#data
<head></head>\x00
#errors
(1,6): expected-doctype-but-got-start-tag
(1,14): unexpected-null-character

#document
| <html>
|   <head>
|   <body>


#data
<table>\x0c</table>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <table>
|       " "


#data
<select></regularitem>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,22): unexpected-end-tag
(1,22): expected-closing-tag-but-got-eof

#document
| <html>
|   <head>
|   <body>
|     <select>


#data
<div><select></div>text
#errors
(1,1): expected-doctype-but-got-start-tag
(1,19): unexpected-end-tag-in-select
(1,23): expected-closing-tag-but-got-eof

#document
| <html>
|   <head>
|   <body>
|     <div>
|       <select>
|         "text"


#data
<figure><frameset><h1><aside><figcaption><html><frameset>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,22): unexpected-token-in-frameset
(1,29): unexpected-token-in-frameset
(1,41): unexpected-token-in-frameset
(1,47): unexpected-start-tag
(1,57): unexpected-start-tag-ignored
(1,57): expected-closing-tag-but-got-eof

#document
| <html>
|   <head>
|   <frameset>


#script-off
#data
<head><noscript><!DOCTYPE html></noscript>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,31): unexpected-doctype

#script-off
#document
| <html>
|   <head>
|     <noscript>
|   <body>


#data
<head><noscript></p></noscript>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,20): unexpected-end-tag

#script-off
#document
| <html>
|   <head>
|     <noscript>
|   <body>
